feat: Enhance ChemSelect with HSP, RDKit and dark mode. New brand HSPmol
Integrates RDKit.js for cheminformatics capabilities, allowing for more advanced molecular analysis.
Adds a dark mode toggle to the sidebar and applies it globally for improved user experience and readability.
Updates the UI for better contrast and visual appeal in both light and dark themes.
Introduces Hansen Solubility Parameters (HSP) to the `Molecule` type and adds them as a new default axis in the dashboard for comparative analysis.
Renames the application title in `index.html` from "ChemSelect" to "HSPmol".